requirements.txt
absl-py==2.3.1
attrs==25.4.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dnspython==2.8.0
eventlet==0.40.3
Flask==3.1.2
Flask-Login==0.6.3
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
flatbuffers==25.9.23
fonttools==4.60.1
greenlet==3.2.4
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
jax==0.6.2
jaxlib==0.6.2
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
mediapipe==0.10.14
ml_dtypes==0.5.3
MouseInfo==0.1.3
mysqlclient==2.2.7
numpy==2.2.6
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opt_einsum==3.4.0
packaging==25.0
pillow==12.0.0
protobuf==4.25.8
PyAutoGUI==0.9.54
pycparser==2.23
PyGetWindow==0.0.9
PyMsgBox==2.0.1
PyMySQL==1.1.2
pyparsing==3.2.5
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-engineio==4.12.3
python-socketio==5.14.2
pytweening==1.2.0
requests==2.32.5
scipy==1.15.3
simple-websocket==1.1.0
six==1.17.0
sounddevice==0.5.3
SQLAlchemy==2.0.44
typing_extensions==4.15.0
urllib3==2.5.0
websocket-client==1.9.0
Werkzeug==3.1.3
wsproto==1.2.0
